app/requirements.txt
redis
django
Celery
djangorestframework
psycopg2
requests
